requirements.txt
django==5.1.6 # frameworknya
psycopg2-binary==2.9.10 # driver postgres
pillow==11.1.0 # untuk mengolah gambar
django-ninja==1.3.0
django-ninja-simple-jwt==0.6.1
locust==2.32.10